requirements/dev.txt
django-nose
coverage
